long to int

+++ fr.ensisa.hassenforder.transportation.empty/fr.ensisa.hassenforder.transportation.server.empty/src/fr/ensisa/hassenforder/transportation/server/network/TerminalReader.java
@@ -39,7 +39,7 @@ public class TerminalReader extends BasicAbstractReader {
     }
 
     private void readUseTicket() {
-        idPass = readInt();
+        idPass = readLong();
         idTicket = readString();
         count = readInt();
     }
